+++ lib/vcs/src/vcs.ml
@@ -1,10 +1,22 @@
+(*******************************************************************************)
+(*  Vcs - a versatile OCaml library for Git interaction                        *)
+(*  Copyright (C) 2024 Mathieu Barbin                                          *)
+(*                                                                             *)
+(*  This file is part of Vcs.                                                  *)
+(*                                                                             *)
+(*  Vcs is free software; you can redistribute it and/or modify it under       *)
+(*  the terms of the GNU Lesser General Public License as published by the     *)
+(*  Free Software Foundation either version 3 of the License, or any later     *)
+(*  version, with the LGPL-3.0 Linking Exception.                              *)
+(*                                                                             *)
+(*  Vcs is distributed in the hope that it will be useful, but WITHOUT ANY     *)
+(*  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS  *)
+(*  FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License and    *)
+(*  the file `NOTICE.md` at the root of this repository for more details.      *)
+(*                                                                             *)
+(*  You should have received a copy of the GNU Lesser General Public License   *)
+(*  and the LGPL-3.0 Linking Exception along with this library. If not, see    *)
+(*  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.       *)
+(*******************************************************************************)
+
 let hello_world = [%sexp "Hello, World!"]
-
-let print_cmd =
-  Command.basic
-    ~summary:"print hello world"
-    (let%map_open.Command () = return () in
-     fun () -> print_s hello_world)
-;;
-
-let main = Command.group ~summary:"" [ "print", print_cmd ]