+++ lib/vcs_git/test/test__vcs_git.ml
@@ -18,8 +18,3 @@
 (*  and the LGPL-3.0 Linking Exception along with this library. If not, see    *)
 (*  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.       *)
 (*******************************************************************************)
-
-let%expect_test "hello" =
-  print_s Vcs.hello_world;
-  [%expect {| "Hello, World!" |}]
-;;