+++ lib/vcs_param/src/vcs_param.mli
@@ -18,5 +18,3 @@
 (*_  and the LGPL-3.0 Linking Exception along with this library. If not, see    *)
 (*_  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.       *)
 (*_******************************************************************************)
-
-(*_ This signature is deliberately empty. *)