+++ CHANGES.md
@@ -1,6 +1,12 @@
-## unreleased
+## Unreleased
 
 ### Added
+
+- Add license and notices.
+- Add libraries skeletons with their opam files.
+
 ### Changed
+
 ### Fixed
+
 ### Removed